third_party/WebKit/LayoutTests/fast/dom/shadow/get-destination-insertion-points-expected.txt

Tests for getDestinationInsertionPoints().

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
PASS nodeList.length is expectedNodes.length
PASS nodeList.item(0) is expectedNodes[0]
Only unclosed insertion points should be returned for getDestinationInsertionPoints().
PASS document.getElementById("child-3").getDestinationInsertionPoints().length is 0
PASS successfullyParsed is true

TEST COMPLETE
